+++ Controllers/DreamTeamController.cs
@@ -111,6 +111,10 @@ namespace FutManager.Controllers
 
         public IActionResult Details(int id)
         {
+            if (!DataService.GetDreamTeams().Any(x => x.Id == id))
+            {
+                return RedirectToAction(actionName: "Index", controllerName: "Error");
+            }
             List<DreamTeam> dreamTeams = new List<DreamTeam>(DataService.GetDreamTeams());
             DreamTeam x = dreamTeams.FirstOrDefault(y => id == y.Id);
             List<Player> players = new List<Player>(DataService.GetPlayers());
@@ -129,6 +133,10 @@ namespace FutManager.Controllers
 
         public RedirectToActionResult Add(string name, string creator, int GoalkeeperId, int LeftDefenderId, int RightDefenderId, int LeftMidfielderId, int RightMidfielderId, int LeftForwardId, int RightForwardId, int ManagerId)
         {
+            if (name == null || creator == null || !DataService.GetPlayers().Any(x => x.Id == GoalkeeperId) || !DataService.GetPlayers().Any(x => x.Id == LeftDefenderId) || !DataService.GetPlayers().Any(x => x.Id == RightDefenderId) || !DataService.GetPlayers().Any(x => x.Id == LeftMidfielderId) || !DataService.GetPlayers().Any(x => x.Id == RightMidfielderId) || !DataService.GetPlayers().Any(x => x.Id == LeftForwardId) || !DataService.GetPlayers().Any(x => x.Id == RightForwardId) || !DataService.GetManagers().Any(x => x.Id == ManagerId))
+            {
+                return RedirectToAction(actionName: "Index", controllerName: "Error");
+            }
             List<int> ids = new List<int>();
             ids.Add(GoalkeeperId);
             ids.Add(LeftDefenderId);
@@ -148,10 +156,18 @@ namespace FutManager.Controllers
 
         public IActionResult Delete(int id)
         {
+            if (!DataService.GetDreamTeams().Any(x => x.Id == id))
+            {
+                return RedirectToAction(actionName: "Index", controllerName: "Error");
+            }
             return View(id);
         }
         public RedirectToActionResult Remove(int id, string password)
         {
+            if (!DataService.GetDreamTeams().Any(x => x.Id == id))
+            {
+                return RedirectToAction(actionName: "Index", controllerName: "Error");
+            }
             if (password == "password")
             {
                 DataService.DeleteDreamteam(id);
